Separates data handler in two files.
